create separate component for ContributorAvatar
